+++ src/main/groovy/org/jetbrains/intellij/dependency/CustomPluginsRepository.groovy
@@ -1,5 +1,6 @@
 package org.jetbrains.intellij.dependency
 
+import kotlin.text.Charsets
 import org.gradle.api.Project
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.annotations.Nullable
@@ -7,6 +8,7 @@ import org.jetbrains.intellij.Utils
 
 import java.nio.file.Files
 import java.nio.file.Paths
+import java.security.MessageDigest
 
 class CustomPluginsRepository implements PluginsRepository {
 
@@ -52,7 +54,8 @@ class CustomPluginsRepository implements PluginsRepository {
         // todo: a better way to define cache directory
         String gradleHomePath = project.gradle.gradleUserHomeDir.absolutePath
         String mavenCacheDirectoryPath = Paths.get(gradleHomePath, 'caches/modules-2/files-2.1').toString()
-        return Paths.get(mavenCacheDirectoryPath, 'com.jetbrains.intellij.idea').toString()
+        String hash = new String(MessageDigest.getInstance("SHA-1").digest(repoUrl.getBytes(Charsets.UTF_8)))
+        return Paths.get(mavenCacheDirectoryPath, "com.jetbrains.intellij.idea", hash).toString()
     }
 
     @NotNull