+++ CHANGES.md
@@ -5,6 +5,7 @@
 - migrate from bintray [#594](../../issues/594)
 - exclude kotlin-reflect and kotlin-text from the runtime if kotlin is used in plugin [#585](../../issues/585)
 - respect overridden `build` directory [#602](../../issues/602)
+- store cache of plugins from different custom repositories in different directories [#579](../../issues/579)
 
 ## 0.7.2